feat: Missing files (remove tsnode)
