+++ package.json
@@ -1,8 +1,12 @@
 {
   "name": "radiant-backend",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "dist/index.js",
   "license": "MIT",
+  "scripts": {
+    "build": "tsc",
+    "start": "node ."
+  },
   "dependencies": {
     "@prisma/client": "^3.11.1",
     "@types/node-fetch": "^2.6.1",
@@ -24,7 +28,6 @@
     "@types/express": "^4.17.13",
     "@types/node": "^17.0.23",
     "nodemon": "^2.0.15",
-    "ts-node": "^10.7.0",
     "typescript": "^4.6.3"
   }
 }
